client: this refactor focuses on improving the readability and structural clarity
